Revert "test: move execution catalog proof to owner boundary"
This reverts commit 72cab3f222d78190001d13d602e376d465094c08.
